WebPageFactory implemented with cglib; WebChildElement interfaces refactoring; JsOperationResult refactoring; WebPageService implemented; Assert functions added; InvocationName added to element actions;

+++ environment/src/main/java/io/perfeccionista/framework/plugin/Color.java
@@ -11,6 +11,6 @@ public interface Color {
 
     int getB();
 
-    int getAlpha();
+    double getAlpha();
 
 }
